+++ src/rpc/mod.rs
@@ -103,6 +103,7 @@ macro_rules! for_each_method {
         $callback!(crate::rpc::eth::EthNewBlockFilter);
         $callback!(crate::rpc::eth::EthUninstallFilter);
         $callback!(crate::rpc::eth::EthSyncing);
+        $callback!(crate::rpc::eth::EthTraceBlock);
         $callback!(crate::rpc::eth::Web3ClientVersion);
         $callback!(crate::rpc::eth::EthSendRawTransaction);